refactor(tests): restructure test suite into unit and integration
Reorganizes the entire /tests directory to establish a clear and conventional separation between unit and integration tests.

- All unit tests (for services, middlewares, etc.) are now located under `tests/unit/` and grouped by module.
- All integration/e2e tests (for routes) are now located under `tests/integration/`.
- Old, flat test directories (`tests/auth/`, `tests/middlewares/`) have been removed.

This new structure is cleaner, more scalable, and aligns with standard testing practices.
